requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
PyMuPDF>=1.23.0
onnxruntime-directml>=1.17.0
pyyaml>=6.0
Pillow>=10.0.0
plotly>=5.18.0
lxml>=4.9.0
html5lib>=1.1
openpyxl>=3.1.0
chardet>=5.2.0
